+++ CHANGELOG.md
@@ -7,6 +7,10 @@ release date when you use `npm version` (see `README.md`).
 
 ## [Unreleased]
 
+### Added
+
+- Add a timestamp to messages.
+
 ## [0.15.1][] - 2022-07-04
 
 ### Fixed
@@ -371,7 +375,5 @@ release date when you use `npm version` (see `README.md`).
 [0.14.2]: https://github.com/webxdc/webxdc-dev/tree/v0.14.2
 [unreleased]: https://github.com/webxdc/webxdc-dev/compare/v0.15.0...HEAD
 [0.15.0]: https://github.com/webxdc/webxdc-dev/tree/v0.15.0
-
-
-[Unreleased]: https://github.com/webxdc/webxdc-dev/compare/v0.15.1...HEAD
+[unreleased]: https://github.com/webxdc/webxdc-dev/compare/v0.15.1...HEAD
 [0.15.1]: https://github.com/webxdc/webxdc-dev/tree/v0.15.1